+++ apps/web/api/reservation.ts
@@ -1,6 +1,42 @@
 import { API_ENDPOINTS } from "@repo/constants";
-import { type ISeat, type IReservation, type ReservationRequestBody } from "@repo/types";
+import { type ReservationRequestBody, type IReservation } from "@repo/types";
 import { axiosRequester } from "@/lib/axios";
+
+interface GetDashboardProps {
+  userId: string;
+}
+/**
+ * 대시보드 페이지를 조회하는 API 함수입니다.
+ * @returns IReservation 배열을 반환합니다.
+ */
+export const getDashboard = async ({ userId }: GetDashboardProps): Promise<IReservation[]> => {
+  const { data } = await axiosRequester<IReservation[]>({
+    options: {
+      method: "GET",
+      url: API_ENDPOINTS.RESERVATION.GET_USER_RESERVATIONS(userId),
+    },
+  });
+
+  return data;
+};
+
+/**
+ * 대시보드 페이지의 회의를 종료하는 API 함수입니다.
+ * @returns IReservation 객체를 반환합니다.
+ */
+export const patchMeetingStatus = async (_id: string): Promise<IReservation> => {
+  const { data } = await axiosRequester<IReservation>({
+    options: {
+      method: "PATCH",
+      url: API_ENDPOINTS.RESERVATION.UPDATE_RESERVATION(_id),
+      data: {
+        status: "completed",
+      },
+    },
+  });
+
+  return data;
+};
 
 /**
  * 좌석예약 페이지를 조회하는 API 함수입니다.
@@ -11,17 +47,6 @@ export const getSeats = async (date: string): Promise<IReservation[]> => {
     options: {
       method: "GET",
       url: API_ENDPOINTS.RESERVATION.GET_RESERVATIONS_BY_TYPE_AND_DATE("seat", date),
-    },
-  });
-
-  return data;
-};
-
-export const getAllSeats = async (): Promise<ISeat[]> => {
-  const { data } = await axiosRequester<ISeat[]>({
-    options: {
-      method: "GET",
-      url: API_ENDPOINTS.ITEMS.GET_ALL("seat"),
     },
   });
